feat : map generator

+++ ProjectAM/Assets/Scripts/ProceduralGeneration/Map/MapNode.cs
@@ -2,7 +2,7 @@ public class MapNode
 {
     public int row;
     public int col;
-    public int perlin;
+    public float perlin;
     public int areaID;
     public char type;
 }
